refactor: syncRecentWindow bridge takes sinceMillis instead of daysBack
Mirrors the native SDK change: pass an absolute since-timestamp (ms) so the app
can re-read from (lastSyncedAt - overlap) rather than a fixed day count.

+++ android/src/main/java/expo/modules/openwearables/OpenWearablesModule.kt
@@ -112,7 +112,7 @@ public class OpenWearablesModule : Module() {
       OpenWearablesHealthSDK.getInstance().syncNow()
     }
 
-    AsyncFunction("syncRecentWindow") Coroutine { _: Int? ->
+    AsyncFunction("syncRecentWindow") Coroutine { _: Double? ->
       // iOS-only catch-up sync; no-op on Android.
       return@Coroutine true
     }
